test(nestjs-signed-url): move fixtures out of src
