+++ packages/nestjs-signed-url/src/gcs/tests/gateway.test.ts
@@ -1,4 +1,4 @@
-import type { FakeGcsClient }   from '../../../tests/gcs.client.interfaces.js'
+import type { FakeGcsClient }   from '../../../tests/fixtures/gcs.client.interfaces.js'
 
 import assert                   from 'node:assert/strict'
 import { beforeEach }           from 'node:test'
@@ -7,7 +7,7 @@ import { it }                   from 'node:test'
 import { mock }                 from 'node:test'
 
 import { GcsSignedUrlGateway }  from '../gateway.js'
-import { createFakeGcsStorage } from '../../../tests/gcs.client.fixture.js'
+import { createFakeGcsStorage } from '../../../tests/fixtures/gcs.client.fixture.js'
 
 describe('GcsSignedUrlGateway', () => {
   let gateway: GcsSignedUrlGateway
