+++ packages/nestjs-signed-url/tests/fixtures/gcs.client.fixture.ts
