+++ packages/nestjs-signed-url/tests/fixtures/gcs.client.interfaces.ts
