+++ packages/nestjs-signed-url/tests/fixtures/gcs.module.interfaces.ts
