+++ packages/nestjs-signed-url/tests/fixtures/signer.fixture.ts
@@ -1,6 +1,6 @@
-import type { SignedUrlReadOptions }      from '../interfaces.js'
-import type { SignedUrlWriteOptions }     from '../interfaces.js'
-import type { SignedUrl }                 from '../interfaces.js'
+import type { SignedUrlReadOptions }      from '../../src/interfaces.js'
+import type { SignedUrlWriteOptions }     from '../../src/interfaces.js'
+import type { SignedUrl }                 from '../../src/interfaces.js'
 import type { ReadCall }                  from './signer.interfaces.js'
 import type { TestingSignedUrlProvider }  from './signer.interfaces.js'
 import type { WriteCall }                 from './signer.interfaces.js'
